refactor: create class for localized web content
